move federator_info stuff
